LayoutTests/imported/w3c/web-platform-tests/webcodecs/full-cycle-test.https.any_h264_avc-expected.txt


Harness Error (FAIL), message = TypeError: undefined is not an object (evaluating 'encoder_color_space.primaries')

FAIL Encoding and decoding cycle assert_equals: frames_decoded expected 16 but got 0
FAIL Encoding and decoding cycle w/ stripped color space assert_equals: frames_decoded expected 16 but got 0
